+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - New program:      eclipseFactor2GriddedData.cpp: returns a global grid with eclipse factors.
 - New program:      orbit2EarhFixedOrbit: Rotate an orbit into a rotation earth fixed frame.
 - New program:      GriddedDataTimeSeries2PotentialCoefficients.
+- New program:      SinexMetadata2GlonassFrequencyNumber.
 - New class:        In Gravityfield/Tides/MiscAccelerations: Group.
 - New option:       GriddedData2PotentialCoefficients: Handle multiple input data columns at once.
 - New option:       GriddedDataReduceSampling: tesseroid volume conserving for digital elevation models.